Refactored env and hashes

+++ srcs/lib/libft/srcs/ft_strsub.c
@@ -6,13 +6,13 @@
 /*   By: bwerewol <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/20 17:45:37 by bwerewol          #+#    #+#             */
-/*   Updated: 2019/02/19 13:01:12 by bwerewol         ###   ########.fr       */
+/*   Updated: 2019/02/19 18:23:33 by nkertzma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strsub(char const *s, unsigned int start, size_t len)
+char	*ft_strsub(char const *s, size_t start, size_t len)
 {
 	char	*str;
